Added Sections (Resources 1-3)
I added the html for the resources. I put the style (id) in the html at the bottom since the css file is being worked on separately. I also linked the resource pages 1-3 to the courseOne page. (I also added the style.css file to this branch as a placeholder since the navbar needs it)